fix all

+++ task1/triangle_square.rb
@@ -1,5 +1,5 @@
 puts "Основание треугольника?"
-length  = gets.chomp.to_i
+length  = gets.chomp.to_f
 puts "Высота?"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 puts "Площадь треугольника #{ 0.5 * length * height}"
